Update cardano-monetary-policy.md
There must be some sort of typo with the last sentence. If the treasury does actually get 1.9 billion ADA every epoch that would be a problem. Just a guess, but maybe you meant 1.9 billion over the 5 years? I was forced to make an edit in order to start this conversation, so it may also be incorrect and I would refer to the experts on what the correction should be.