Use main sphinx_rtd_theme instead of fork
There was no need to use a fork here simply set `html_show_sphinx = 
False`, this removes the need for 
https://github.com/input-output-hk/sphinx_rtd_theme (forked and modified 
with 
https://github.com/input-output-hk/sphinx_rtd_theme/commit/12c5bccf1cc1f5794f482df79d6857e0ec52803f) 
and removes the extra maintence cost.

This change also bumbs the theme version to 0.5.2 which includes some 
fixes to fix compatibility with newer sphinx versions.

FWIW I am the maintainer of sphinx_rtd_theme